+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kewl-e-com",
-  "version": "2.3.1",
+  "version": "2.3.2",
   "private": true,
   "scripts": {
     "dev": "next dev",
