+++ src/app/globals.css
@@ -9,6 +9,7 @@
   --primary2: #00d1ff;
   --secondary1: #ff6b35;
   --secondary2: #9c88ff;
+  --neutral: #0e0e0e;
   --neutral1: #222222;
   --neutral2: #333333;
 }
@@ -34,11 +35,12 @@ body {
   border-color: white;
 }
 
-.css-1tktgsa-MuiPaper-root-MuiPopover-paper-MuiMenu-paper {
-  background-color: var(--neutral1);
+#menu-appbar .MuiPaper-root {
+  background-color: var(--neutral);
+  border-radius: 1rem;
 }
 
 .cl-userProfile-root {
   margin: auto;
-  padding: 2rem;
+  padding-top: 2rem;
 }
